requirements.txt
uvicorn
fastapi
ollama
sqlalchemy
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-dotenv
alembic
authlib
python-jose[cryptography]
sendgrid
openai
pandas
prophet
tavily-python
alpha-vantage
pinecone
sentence-transformers
psycopg2-binary 
newsapi-python